feat: refactor options page into multiple files
- Split monolithic options page into separate components
- Improve code organization and maintainability
- Prepare for future enhancements to options functionality
