+++ pages/history.html
@@ -1,0 +1,15 @@
+<div id="history-page">
+    <h1>Summary History</h1>
+    <table id="history-table">
+        <thead>
+            <tr>
+                <th>URL</th>
+                <th>Title</th>
+                <th>Summarized Time</th>
+                <th>Summary</th>
+                <th>Actions</th>
+            </tr>
+        </thead>
+        <tbody></tbody>
+    </table>
+</div>
